Rename helper_functions module to utils
